Restructure Java cron example to single class and rename to CronJobInitiator and CronJob
