+++ java/patterns-use-cases/README.md
@@ -11,7 +11,7 @@
 - **[Payment State Machines (Advanced)](README.md#payment-state-machines)**: State machine example that tracks a payment process, ensuring consistent processing and cancellations. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/statemachinepayments/PaymentProcessor.java)
 
 #### Scheduling
-- **[Cron Jobs](README.md#cron-jobs)**: Implement a cron service that executes tasks based on a cron expression. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/cron/CronJob.java)
+- **[Cron Jobs](README.md#cron-jobs)**: Implement a cron service that executes tasks based on a cron expression. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/cron/Cron.java)
 - **[Scheduling Tasks](README.md#scheduling-tasks)**: Restate as scheduler: Schedule tasks for later and ensure the task is triggered and executed. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/schedulingtasks/PaymentTracker.java)
 - **[Parallelizing Work](README.md#parallelizing-work)**: Execute a list of tasks in parallel and then gather their result. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/parallelizework/FanOutWorker.java)
 - **[Payment Signals (Advanced)](README.md#payment-signals)**: Handling async payment callbacks for slow payments, with Stripe. [<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/play-button.svg" width="16" height="16">](src/main/java/my/example/signalspayments/PaymentService.java)
@@ -216,7 +216,7 @@ dev.restate.sdk.common.TerminalException: [👻 SIMULATED] This hotel is fully b
 </details>
 
 ## Cron Jobs
-[<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/show-code.svg">](src/main/java/my/example/cron/CronJob.java)
+[<img src="https://raw.githubusercontent.com/restatedev/img/refs/heads/main/show-code.svg">](src/main/java/my/example/cron/Cron.java)
 
 Restate has no built-in functionality for cron jobs.
 But Restate's durable building blocks make it easy to implement a service that does this for us.
@@ -249,7 +249,7 @@ Note that this implementation is fully resilient, but you might need to make som
    ```
 2. Start the cron service and the task service:
    ```shell
-   ./gradlew -PmainClass=my.example.cron.AppMain run
+   ./gradlew -PmainClass=my.example.cron.Cron run
    ```
 3. Register the services (with `--force` to override the endpoint during **development**): 
    ```shell
@@ -258,7 +258,7 @@ Note that this implementation is fully resilient, but you might need to make som
 Send a request to create a cron job that runs every minute:
 
 ```shell
-curl localhost:8080/CronService/create --json '{ 
+curl localhost:8080/CronJobInitiator/create --json '{ 
       "expr": "* * * * *", 
       "service": "TaskService", 
       "method": "executeTask", 
@@ -269,7 +269,7 @@ curl localhost:8080/CronService/create --json '{
 Or create a cron job that runs at midnight:
 
 ```shell
-curl localhost:8080/CronService/create --json '{ 
+curl localhost:8080/CronJobInitiator/create --json '{ 
       "expr": "0 0 * * *", 
       "service": "TaskService", 
       "method": "executeTask",
